0000059: Contacts module doest not show all contacts
Fixed this bug, the math was preventing the last section of up to 3 contacts (the maximum integer remainder for a division by 4 which is used for the contacts list grid) from displaying on the contacts list.
Too much talk for a simple change from floor to ceil, it was faster to fix the bug than type all this stuff... so I took the time to also fix a display problem when the names where too big and they blew horizontally the contacts info bellow the contacts name.

git-svn-id: https://web2project.svn.sourceforge.net/svnroot/web2project@268 cef3e9b6-683d-0410-b475-c2bd3a4d0f1e