Fixed an issue with the projects list derived from issue #220: http://bugs.web2project.net/view.php?id=220
Consisting of a mess with an in_array function and the new $projectArray support variable.
Former lines 261 and 262 are the focus of this fix.

If $projects['project_id'] was not defined when doing...

$projectArray[$project['project_id']] = $project['project_id'];

...it would cause the projectArray to be filled with keys without any proper project_id.
This would cause the in_array check to throw out all projects with project ids within those loose keys instead of the proper already displayed projects ids.

So now we fill the $projectArray array with existing project ids and synchronize the array value with its key.

Then we can use isset instead of in_array for performance reasons.


git-svn-id: https://web2project.svn.sourceforge.net/svnroot/web2project@587 cef3e9b6-683d-0410-b475-c2bd3a4d0f1e