Fix Bug #0000241: Creating/Editing a task in IE 8 does not work http://bugs.web2project.net/view.php?id=241
When adding or editing a task, all the tabs are glued together at submit time with the assistance of some DHTML/Javascript.
If the gluing doesn't work out, only the first fields of the task record are stored, and all the info from the tabs is sent to oblivion.
Restoring these records would require to at least getting the task_owner field populated, or the task becomes phantom (meaning, it is on the DB but the retrieval query misses them because a task should have a owner).
That should be dealt with on another commit.

Since IE 8 behaves more like gecko, we need to check for IEs version number and use the same algorithms as gecko (firefox) and opera.

git-svn-id: https://web2project.svn.sourceforge.net/svnroot/web2project@598 cef3e9b6-683d-0410-b475-c2bd3a4d0f1e