Added more info to our Performance/Debug box so we get more environmental information from the users when debugging, yet non-intrusive: 1) DB Provider description 2) Web Server 3) Server protocol and gateway interface 4) Client Browser 5) URL (Query portion only) 6) Execution time on the server and servers timezone
Hopefuly this will cut some time waiting for feedback.

For those that do not know how to see the Performance box, and want to, you must edit the base.php file and change the line that says:

define('W2P_PERFORMANCE_DEBUG', false);

to

define('W2P_PERFORMANCE_DEBUG', true);

It will show up on the browser at the bottom of any web2Project interface.
Change it back to what it was and it will be switched off.

git-svn-id: https://web2project.svn.sourceforge.net/svnroot/web2project/trunk@602 cef3e9b6-683d-0410-b475-c2bd3a4d0f1e