This commit has 1 important security fix and 1 important performance improvement:
Security Fix:
Tasks on tasks lists and on the calendar were not being checked for Task Access limitations.
Users would not be able to "view" or "edit" them (getting an access denied screen), but they would be able to see them on a listing context.

Performance improvement:
The calendar module got a kick on its logics, and therefore should be 80% faster now.
Also filtered out the Tasks from the inactive projects and template projects from the calendars.

git-svn-id: https://web2project.svn.sourceforge.net/svnroot/web2project@604 cef3e9b6-683d-0410-b475-c2bd3a4d0f1e