0000263: Task I have created not returned correctly Description: I create a task, setting an owner When I am in Tals module, filtering my task by "Task I have created" , there is nothing. Looking in MySQL base, task creator is correctly set. It seems that the request sent selecting "All Tasks I Have Created" is not checking the task creator but task owner. ----
Added the "Tasks That I Am Owner" option on the Filters.
And made it work like the wording says.
So "Tasks That I Am Owner" points to tasks a user is owner of, and "All Tasks I Have Created" points to tasks a user has indeed created.

Added the necessary string for translation.

git-svn-id: https://web2project.svn.sourceforge.net/svnroot/web2project@668 cef3e9b6-683d-0410-b475-c2bd3a4d0f1e