Perms getAllowedItems and getDeniedItems were being called with a table name as first parameter instead of a module name. That would make things being naturally denied for all modules where the module name is different of the table name.
So I made the CW2pObject call the permissions getAllowedItems and getDeniedItems with a module name, as it should, making sure that when the table and the module name match the call is still done correctly.

Modules where the table name and the module name are different must use the third parameter (the module directory name) on their __construct() for the permissions to work correctly.
Because if not, the permissions will be checked against something that is not defined on the gacl tables, and therefore... denied.

Example:

	public function __construct()
	{
		parent::__construct('notes', 'note_id', 'notebook');
	}

In this case the table is 'notes' but the permissions that will be looked up on the gacl tables is for the 'notebook' module.
If the third parameter was not set, then the permissions engine would look for 'notes' on the gacl_axo table, and since it is not there, the permissions engine would deny any action.

git-svn-id: https://web2project.svn.sourceforge.net/svnroot/web2project/trunk@1294 cef3e9b6-683d-0410-b475-c2bd3a4d0f1e