http://bugs.web2project.net/view.php?id=549
Opted for the 1st solution:
"Add any known modules string keys to /locales/en/common.inc
PROS: Faster fix, better performance by not loading every single .inc file into w2P
CONS: The fix is fast because we only need to add modules name strings to one file, but to have it really fixed we'll need to wait for the community to provide translated .inc files before the users see the strings correctly translated on their interface."

Also added all knowned to exist add-on modules (the ones that are on web2project-mod SVN repository) strings.

Community: go ahead and translate the common.inc file on your native language and share it with us on the forums. We'll add it to core.

git-svn-id: https://web2project.svn.sourceforge.net/svnroot/web2project/trunk@1327 cef3e9b6-683d-0410-b475-c2bd3a4d0f1e