Bug 0000567: Email Notification not working when uploading/editing files http://bugs.web2project.net/view.php?id=567
Fixed email notification for files.
Bear in mind that notifications will only be sent to:
Assignees or (if they don't exist) to the project owner.
Besides that also project contacts will bee notified.
Last but not least, if you are an assignee or (if those don't exist) the project owner, and you are doing the file upload then you are not going to get an email, because we presume that you don't need to be notified of what you are doing, or just done.

Also fixed the editing of Files if no file is to sent on the edit process, and you only want to "touch" the file info details.
If no file is being uploaded then we don't need to index the file, because there is nothing new to index anyway.
(As a matter of fact this would lead to an error because we have no access to the temp file for index when editing anymore)

git-svn-id: https://web2project.svn.sourceforge.net/svnroot/web2project@1345 cef3e9b6-683d-0410-b475-c2bd3a4d0f1e