Fixed LDAP authenticator.
SQL fallback is now the last thing to be considered. When trying to login for the first time, if the user authenticates against the LDAP, a new user is created in the database. Each time a LDAP user logs in, the password is compared with the existing one in the database (just to keep things synched).