Changes on the behaviour of the calendar's views.
- On 'Day View' the minicals show only tasks and events associated with the
  current user, no longer showing tasks associated with others, from
  projects belonging to the current user.
- On 'Day View' the 'Tasks assigned to others' list shows tasks that are
  assigned to others but are from projects owned by the current user and
  that are late.
- The other calendar views no longer show tasks assigned to other users, as
  in 'Day View'.

I think this behaviour is more useful (a project manager can get a quick
sense of the state its projects are in right in 'Day View').

This patch also changes the way that task filtering by due date and
completion state on the 'Day View' is done, by bringing it back to the SQL
query. For that purpose code that set task's end dates, if unset, has been
taken out as it's now impossible to create a task without dates.