Fix to control the visibility of the edit, checkin/out and delete icons on the files list, depending on the item permissions and owner.
Now a file can be edited by its owner or by anyone who is given edit
permissions on the file, while a file can be deleted by its owner or by
anyone with delete permissions on it.
These permissions do not depend on the project, task or whatever.

Also included is a fix for the 'Selection' column.