Fixes on the forum messages views:
- The parent message to be passed into the post message form is now
properly computed; with the old code posting a reply to a single message on
a topic (not the top one) would setup the old message as the parent of the
new, instead of using the first in the topic.
- Cosmetic changes on the 'Collapsed' and 'Single Message at a time' views;
now the 'last edited by' stuff is pushed to the right, to use more width
instead of height.
- Cosmetic changes on the 'Collapsed' and 'Single Message at a time' views;
now the message body is shown as written (with newlines instead of running
text).
- Cosmetic change in the 'Collapsed' view, by adding some margin on top to
separate the message body from the rest.