deprecated all the clash functionality as we no longer support it
The calendaring system only provides read-only access to third party
systems (via ical) so it is not the system of record. Therefore, we
can’t base clash detection on it.. in order to do it effectively, we
also need to be able to read from external systems. A CalDav
implementation would likely allow this.